Create caesar.c
This question required me to write a program called caesar that enables you to encrypt messages using Caesar’s cipher, which involves the user providing a key with which we encrypt a certain "Plain text" that the used provides, into a cipher text. This program was also useful in understanding how to set conditions on Command Line Arguments which are relevant to the problem.